--- scripts/app/configs/webpack/production/client/output.ts
@@ -1,18 +1,0 @@
-import path from "path";
-import {DevtoolModuleFilenameTemplateInfo} from "webpack";
-import Application from "../../../../Application";
-import files from "../../../files";
-
-const output = {
-    chunkFilename: files.outputPattern.jsChunk,
-    devtoolModuleFilenameTemplate: (info: DevtoolModuleFilenameTemplateInfo) =>
-        path
-            .relative(Application.srcPath, info.absoluteResourcePath)
-            .replace(/\\/g, "/"),
-    filename: files.outputPattern.js,
-    path: `${Application.buildPath}/public`,
-    publicPath: Application.servedPath,
-
-};
-
-export default output;
